Update to CWMS Plot (#53)
* Updates to TSPlot to parse component user input into plotly parameters

* Updates to TSPlot to parse component user input into plotly parameters

* Used auto formatting to match rest of standards in library. A format standard file exists here `docs\.prettierrc` and if using prettier it should target this to match the other files.

Consider adding "format on save" for your "WORKSPACE"

- Updated plot to show SWT's Top of Conservation instead of Spillway for the example
- Fixed bug where if user did not want to show levels it would throw an error for missing `plotParams` properties
- add plotParams to hook dependencies to ensure plot reloads if a user specifies new params

* Updates to CWMSPlot and starting a new CWMSTable

* Updates to Plots and Tables

* Update to CWMS Plots, notes in pull request on 14 Nov

* Update to CWMS Plot

* Addressing items from PR #53

* Added test code to doc page - compiled and tested against lib. More docs needed

* Update package lock for docs

* Started on cwmsplot docs

* Updates to CWMS Plot which should address all the previous comments, except using deepmerge and skeleton, and adding more helpful error messages and/or stack traces in the console log.

* Remove prop docstrings from TSPlot

Outdated -- need to be updated if included

* Apply "prettier" formatting to CWMSPlot

* Fix syntax of ternary assignment

* Prevent slice error when begin/end not provided

* Standardize id-mapping syntax

* Add useEffect dependencies as necessary

Move tsid and level id-mapping into useEffect calls to avoid redundant
dependencies.

* Use Skeleton as default loadingComponent

* Install deepmerge

* Merge user-provided layout with defaultLayout

Rather than overwriting the defaultLayout with a user-provided layout
when available, use a deepmerge to apply any user-provided layout
options on top of the defaultLayout.

* Reincorporate some of @krowvin's error handling

Was overwritten during an earlier code-merge.

* Improve default yaxis layout building

When adding layout yaxis objects based on time series params, do not add
an additional yaxis if the yaxis is already defined in a previous time
series.  Do not double-count the same yaxis when setting grid columns.

* Use one source of truth for yaxis count

Derive grid.rows from count of "yaxis" in defaultLayout keys

* Change timeseriesParams to timeSeries

Restructure prop to be an object with an id and an optional traceOptions
object.

* Account for "y1" in yaxis property for traces

* Support single entries and strings for plot data

* Change locationLevelParams to locationLevels

Implement the {id, traceOptions} format for the locationLevels prop.

* Change layoutParams to layoutOptions

Maintaining consistency with other props

* Update cwmsjs functions to new version

* Update WIP docs to use new CWMSPlot props

* Remove TSPlot from library

Component is mostly redundant given the updates to CWMSPlot.  May return
in a different format down the line if it would be beneficial.

* Clean up and pare down example plot props

Attempt to keep the example code short to avoid documentation bloat.

* Move CWMSPlot docs to an individual page

Making room for additional plot components in the future.

* Add Plots overview page

* Remove custom loadingComponent from CWMSPlot

This is probably something that should be standardized across WM.

* Remove default title from CWMSPlot

May be desired in many cases to have no plot title.

* Overhaul CWMSPlot documentation

---------

Co-authored-by: Charles Graham <charles.r.graham@usace.army.mil>
Co-authored-by: Brandon Kolze <james.b.kolze@usace.army.mil>