Change to using the Timeseries Location Level endpoint. Updates to CWMSPlot code and Docs as necessary. (#124)
Endpoint requires the units of the location level be specified - not just EN or SI but "ft" for example.

Using Timeseries Location Levels includes building array of datetime, level pairs with adjustments for the first and last entries (only) to match the exact range of the timeseries values data. Additional work may be undertaken to adjust the other items in that array to snap to more sensible values instead of the response from CDA which is based upon the timestamp of the request.

Separately, removed a block of code that was commented-out and was intended to handle the page size based on the time range and pathname interval. Now that page size is included as its own parameter, this is no longer necessary.